feature: add Ctrl+O global hotkey to open/init repository
This fix add a global hotkey, Ctrl-O, that allows you to perform actions
that were previously only possible on the Welcome screen, from anywhere.

- Add global Ctrl+O (Cmd+O on macOS) hotkey for opening a repository.
- Move dialog operation logic `OpenLocalRepository()`  from
  WelcomeToolbar to Launcher, and use it by global hotkey and button
  action in WlcomeToolbar.
- Add hotkey text to "Open repository" button tooltip in welcome view.
- Add new hotkey text in Hotkeys dialog as resource (en_US, ja_JP).
