+++ src/Views/Launcher.axaml.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.IO;
+using System.Threading.Tasks;
 
 using Avalonia;
 using Avalonia.Controls;
@@ -7,6 +9,7 @@ using Avalonia.Interactivity;
 using Avalonia.Markup.Xaml.MarkupExtensions;
 using Avalonia.Media;
 using Avalonia.Platform;
+using Avalonia.Platform.Storage;
 using Avalonia.VisualTree;
 
 namespace SourceGit.Views
@@ -104,6 +107,56 @@ namespace SourceGit.Views
                 WindowState = _lastWindowState;
             else
                 Activate();
+        }
+
+        public async Task OpenLocalRepository()
+        {
+            var vm = App.GetLauncher();
+            var activePage = vm.ActivePage;
+            if (activePage == null || !activePage.CanCreatePopup())
+                return;
+
+            var topLevel = TopLevel.GetTopLevel(this);
+            if (topLevel == null)
+                return;
+
+            var preference = ViewModels.Preferences.Instance;
+            var workspace = preference.GetActiveWorkspace();
+            var initDir = workspace.DefaultCloneDir;
+            if (string.IsNullOrEmpty(initDir) || !Directory.Exists(initDir))
+                initDir = preference.GitDefaultCloneDir;
+
+            var options = new FolderPickerOpenOptions() { AllowMultiple = false };
+            if (Directory.Exists(initDir))
+            {
+                var folder = await topLevel.StorageProvider.TryGetFolderFromPathAsync(initDir);
+                options.SuggestedStartLocation = folder;
+            }
+
+            try
+            {
+                var selected = await topLevel.StorageProvider.OpenFolderPickerAsync(options);
+                if (selected.Count == 1)
+                {
+                    var folder = selected[0];
+                    var folderPath = folder is { Path: { IsAbsoluteUri: true } path } ? path.LocalPath : folder?.Path.ToString();
+                    var repoPath = await ViewModels.Welcome.Instance.GetRepositoryRootAsync(folderPath);
+                    if (!string.IsNullOrEmpty(repoPath))
+                    {
+                        await ViewModels.Welcome.Instance.AddRepositoryAsync(repoPath, null, false, true);
+                        ViewModels.Welcome.Instance.Refresh();
+                    }
+                    else if (Directory.Exists(folderPath))
+                    {
+                        var test = await new Commands.QueryRepositoryRootPath(folderPath).GetResultAsync();
+                        ViewModels.Welcome.Instance.InitRepository(folderPath, null, test.StdErr);
+                    }
+                }
+            }
+            catch (Exception exception)
+            {
+                Models.Notification.Send(null, $"Failed to open repository: {exception.Message}", true);
+            }
         }
 
         protected override async void OnOpened(EventArgs e)
@@ -219,6 +272,13 @@ namespace SourceGit.Views
                 if (e.Key == Key.W)
                 {
                     vm.CloseTab(null);
+                    e.Handled = true;
+                    return;
+                }
+
+                if (e.Key == Key.O)
+                {
+                    await OpenLocalRepository();
                     e.Handled = true;
                     return;
                 }
